requirements.txt
numpy==1.25.0
alive-progress==3.1.4
Pillow==9.5.0
transformers==4.30.2
huggingface-hub==0.15.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
ultralytics==8.0.114
tokenizers==0.13.3
gitpython==3.1.30